+++ src/GitHub_Updater/Settings.php
@@ -221,7 +221,7 @@ class Settings extends Base {
 					</div>
 				<?php elseif ( ( isset( $_GET['refresh_transients'] ) && true == $_GET['refresh_transients'] ) ) : ?>
 					<div class="updated">
-						<p><?php esc_html_e( 'Transients refreshed.', 'github-updater' ); ?></p>
+						<p><?php esc_html_e( 'Cache refreshed.', 'github-updater' ); ?></p>
 					</div>
 				<?php endif; ?>
 
@@ -272,7 +272,7 @@ class Settings extends Base {
 					</form>
 					<?php $refresh_transients = add_query_arg( array( 'github_updater_refresh_transients' => true ), $action ); ?>
 					<form class="settings" method="post" action="<?php esc_attr_e( $refresh_transients ); ?>">
-						<?php submit_button( esc_html__( 'Refresh Transients', 'github-updater' ) ); ?>
+						<?php submit_button( esc_html__( 'Refresh Cache', 'github-updater' ) ); ?>
 					</form>
 				<?php endif; ?>
 			<?php endif; ?>